+++ sample-android/src/main/java/com/github/panpf/zoomimage/sample/ui/examples/compose/BaseZoomImageSample.kt
@@ -20,7 +20,6 @@ import com.github.panpf.zoomimage.compose.rememberZoomImageLogger
 import com.github.panpf.zoomimage.compose.rememberZoomState
 import com.github.panpf.zoomimage.compose.zoom.ScrollBarSpec
 import com.github.panpf.zoomimage.compose.zoom.ZoomAnimationSpec
-import com.github.panpf.zoomimage.compose.zoom.heartbeat
 import com.github.panpf.zoomimage.sample.compose.widget.ZoomImageMinimap
 import com.github.panpf.zoomimage.sample.compose.widget.ZoomImageTool
 import com.github.panpf.zoomimage.sample.compose.widget.rememberMyDialogState
@@ -28,7 +27,6 @@ import com.github.panpf.zoomimage.sample.settingsService
 import com.github.panpf.zoomimage.sample.ui.util.compose.valueOf
 import com.github.panpf.zoomimage.subsampling.TileAnimationSpec
 import com.github.panpf.zoomimage.util.Logger
-import com.github.panpf.zoomimage.zoom.OneFingerScaleSpec
 import com.github.panpf.zoomimage.zoom.ReadMode
 import com.github.panpf.zoomimage.zoom.ScalesCalculator
 
@@ -49,7 +47,6 @@ fun BaseZoomImageSample(
     val contentScaleName by settingsService.contentScale.stateFlow.collectAsState()
     val alignmentName by settingsService.alignment.stateFlow.collectAsState()
     val threeStepScale by settingsService.threeStepScale.stateFlow.collectAsState()
-    val oneFingerScale by settingsService.oneFingerScale.stateFlow.collectAsState()
     val rubberBandScale by settingsService.rubberBandScale.stateFlow.collectAsState()
     val readModeEnabled by settingsService.readModeEnabled.stateFlow.collectAsState()
     val readModeAcceptedBoth by settingsService.readModeAcceptedBoth.stateFlow.collectAsState()
@@ -100,10 +97,6 @@ fun BaseZoomImageSample(
     val zoomState = rememberZoomState(rememberZoomImageLogger(level = logLevel)).apply {
         LaunchedEffect(threeStepScale) {
             zoomable.threeStepScale = threeStepScale
-        }
-        LaunchedEffect(oneFingerScale) {
-            zoomable.oneFingerScaleSpec = if (oneFingerScale)
-                OneFingerScaleSpec.heartbeat(zoomable) else null
         }
         LaunchedEffect(rubberBandScale) {
             zoomable.rubberBandScale = rubberBandScale